+++ delivery_roulier/__init__.py
@@ -1,1 +1,3 @@
 from . import models
+
+from .decorator import implemented_by_carrier
