+++ delivery_roulier/__manifest__.py
@@ -4,7 +4,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     'name': 'Delivery Carrier Roulier',
-    'version': '9.0.0.0.0',
+    'version': '10.0.1.0.0',
     'author': 'Akretion,Odoo Community Association (OCA)',
     'summary': 'Integration of multiple carriers',
     'maintainer': 'Akretion, Odoo Community Association (OCA)',
