+++ delivery_roulier/models/stock_picking.py
@@ -4,41 +4,19 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from datetime import datetime, timedelta
-from functools import wraps
 import logging
 
-from openerp import models, fields, api
-from openerp.tools.translate import _
-from openerp.exceptions import Warning as UserError
+from odoo import models, fields, api
+from odoo.tools.translate import _
+from odoo.exceptions import UserError
+
+from ..decorator import implemented_by_carrier
 
 _logger = logging.getLogger(__name__)
 try:
     from roulier import roulier
 except ImportError:
     _logger.debug('Cannot `import roulier`.')
-
-
-def implemented_by_carrier(func):
-    """Decorator: call _carrier_prefixed method instead.
-
-    Usage:
-        @implemented_by_carrier
-        def _do_something()
-        def _laposte_do_something()
-        def _gls_do_something()
-
-    At runtime, picking._do_something() will try to call
-    the carrier spectific method or fallback to generic _do_something
-    """
-    @wraps(func)
-    def wrapper(cls, *args, **kwargs):
-        fun_name = func.__name__
-        fun = '_%s%s' % (cls.carrier_type, fun_name)
-        if not hasattr(cls, fun):
-            fun = '_roulier%s' % (fun_name)
-            # return func(cls, *args, **kwargs)
-        return getattr(cls, fun)(*args, **kwargs)
-    return wrapper
 
 
 class StockPicking(models.Model):
