+++ delivery_roulier/views/stock_quant_package.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
     <!-- add Tracking button -->
     <record id="view_quant_package_form" model="ir.ui.view">
         <field name="model">stock.quant.package</field>
@@ -12,6 +11,4 @@
             </div>
         </field>
     </record>
-
-</data>
-</openerp>
+</odoo>
